requirements.txt
# Incident Response Platform - Combined Phase-1 + Phase-2
# Python dependencies

# Core
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM & AI
openai>=1.0.0

# Data Processing
pyyaml>=6.0.1
networkx>=3.0

# HTTP & API
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
